feature: preserve html inside target elements
previously, all HTML inside the target element was stripped away
when the text was split.

With this feature, nested HTML tags within the target elements
will be preserved.
