+++ .eslintrc.json
@@ -3,6 +3,7 @@
   "extends": ["airbnb-base", "prettier"],
   "parserOptions": { "ecmaVersion": 12, "sourceType": "module" },
   "rules": {
+    "no-lonely-if": 0,
     "no-plusplus": 0,
     "no-param-reassign": 0,
     "no-else-return": 0,
