+++ __tests__/utils/getTextContent.test.js
@@ -1,16 +1,14 @@
 import getTextContent from '../../lib/utils/getTextContent'
 
 describe(`utils.getTextContent(element)`, () => {
-  const textContent = 'Hello Word.'
-  it(`Returns the textContent of element`, () => {
-    const element = document.createElement('div')
-    element.textContent = textContent
-    expect(getTextContent(element)).toEqual(textContent)
+  const textContent = 'Hello World.'
+  it(`Returns the value of a text node`, () => {
+    const node = document.createTextNode(textContent)
+    expect(getTextContent(node)).toEqual(textContent)
   })
 
-  it(`Returns the text content of an element that contains nested HTML`, () => {
-    const element = document.createElement('div')
-    element.innerHTML = `<div>Hello</div> <span>Word</span>.`
-    expect(getTextContent(element)).toEqual(textContent)
+  it(`Removes extra white space`, () => {
+    const node = document.createTextNode('  Hello   World.\n ')
+    expect(getTextContent(node)).toEqual(textContent)
   })
 })
