+++ lib/SplitType.js
@@ -4,6 +4,7 @@ import toArray from './utils/toArray'
 import getTargetElements from './utils/getTargetElements'
 import { Data, RemoveData } from './Data'
 import split from './split'
+import repositionAfterSplit from './repositionAfterSplit'
 import defaults from './defaults'
 
 // Default Settings for `SplitType` instances
@@ -125,11 +126,16 @@ export default class SplitType {
     this.elements.forEach((element) => {
       // Add the split text nodes from this element to the arrays of all split
       // text nodes for this instance.
-      const { lines, words, chars } = split(element, this.settings)
-      this.lines = this.lines.concat(lines)
-      this.words = this.words.concat(words)
-      this.chars = this.chars.concat(chars)
+      Data(element).isRoot = true
+      split(element, this.settings)
       Data(element).isSplit = true
+      const { lines, words, chars } = repositionAfterSplit(
+        element,
+        this.settings
+      )
+      this.lines = [...this.lines, ...lines]
+      this.words = [...this.words, ...words]
+      this.chars = [...this.chars, ...chars]
     })
 
     // Set isSplit to true for the SplitType instance
