+++ lib/utils/createElement.js
@@ -1,5 +1,7 @@
 import forEach from './forEach'
 import isNode from './isNode'
+import isString from './isString'
+import createTextNode from './createTextNode'
 
 /**
  * Create an HTML element with the the given attributes
@@ -36,6 +38,7 @@ export default function createElement(name, attributes) {
     else if (attribute === 'children') {
       forEach(value, (child) => {
         if (isNode(child)) element.appendChild(child)
+        if (isString(child)) element.appendChild(createTextNode(child))
       })
     }
     // Handle standard HTML attributes
