+++ lib/utils/createTextNode.js
@@ -1,0 +1,3 @@
+export default function createTextNode(string) {
+  return document.createTextNode(string)
+}
