+++ lib/utils/getTextContent.js
@@ -1,22 +1,12 @@
 /**
- * Gets the text content of an HTML element.
+ * Gets the text content of a text node.
+ * Trims white space from start/end of string
+ * Removes extra white space
  *
- * Optionally, <br> tags can be replaced with a unique string so they can be
- * converted back HTML later on.
- *
- * @param {HTMLElement} element
- * @param {string} BR_SYMBOL
+ * @param {Node} node a text node
  * @return {string} the text content of the given element
  */
-export default function getTextContent(element, LINE_BREAK_SYMBOL) {
-  const brTag = /<br\s*\/?>/g
-  let textContent = element.textContent
-  if (LINE_BREAK_SYMBOL) {
-    const innerHTML = element.innerHTML
-    const tempDiv = document.createElement('div')
-    tempDiv.innerHTML = innerHTML.replace(brTag, ` ${LINE_BREAK_SYMBOL} `)
-    textContent = tempDiv.textContent
-  }
-  // Remove extra white space
-  return textContent.replace(/\s+/g, ' ').trim()
+export default function getTextContent(node) {
+  const value = node.nodeValue || ''
+  return value.replace(/\s+/g, ' ').trim()
 }
