+++ lib/utils/isNode.js
@@ -4,7 +4,6 @@ import isObject from './isObject'
  * Returns true if `input` is one of the following:
  * - `Element`
  * - `Text`
- * - `Document`
  * - `DocumentFragment`
  */
 export default function isNode(input) {
