+++ package.json
@@ -18,6 +18,7 @@
     "puppeteer:watch": "NODE_ENV=development jest --testMatch='**/__stories__/**/?*.test.js' --env=jsdom  --watchAll",
     "puppeteer:static": "yarn storybook:build && jest --testMatch='**/__stories__/**/?*.test.js' --env=jsdom",
     "build": "rollup -c",
+    "build:watch": "onchange 'lib/**/*.js' -- rollup -c",
     "lint": "./node_modules/.bin/eslint . --color && echo \"eslint: no lint errors\"",
     "storybook": "start-storybook -p 6006 ",
     "storybook:build": "build-storybook"
@@ -56,6 +57,7 @@
     "eslint-plugin-import": "^2.26.0",
     "husky": "^7.0.0",
     "jest": "27.x",
+    "onchange": "^7.1.0",
     "prettier": "^2.6.2",
     "pretty-quick": "^3.1.3",
     "puppeteer": "^13.7.0",
